+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 2.6.1 - 2021-07-01
 
 ### Added
+- Slot carousel header in component (`QasCard`). 
 - Added ref `form` to `q-form` in component (`QasFormView`).
 
 ## 2.6.0 - 2021-06-29